refactor(all)

+++ src/studyCloud/Application/Abstractions/Services/IBlockService.cs
@@ -6,5 +6,5 @@ public interface IBlockService
 {
     Task<bool> Block(Block block);
     Task<bool> Unblock(Block block);
-    Task<List<Block>> GetByUserId(int userId);
+    Task<List<Block>> GetByUserId(long userId);
 }
