+++ src/studyCloud/Application/Abstractions/Services/IFollowService.cs
@@ -5,9 +5,9 @@ namespace Application.Abstractions.Services;
 
 public interface IFollowService
 {
-    Task<Follow> FindByMemberIdAndFollowMemberId(int memberId, int followMemberId);
+    Task<Follow> FindByMemberIdAndFollowMemberId(long memberId, long followMemberId);
     Task<Follow> DeleteAsync(Follow follow);
     Task<Follow> DeleteFollowerAsync(Follow follow);
-    Task<FollowerDto> GetFollowers(int memberId);
-    Task<List<Follow>> GetFollowings(int memberId);
+    Task<FollowerDto> GetFollowers(long memberId);
+    Task<List<Follow>> GetFollowings(long memberId);
 }
