+++ src/studyCloud/Application/Features/Comments/Comment/Commands/UpdateComment/UpdateCommentCommandRequest.cs
@@ -4,9 +4,9 @@ namespace Application.Features.Comments.Comment.Commands.UpdateComment;
 
 public class UpdateCommentCommandRequest : IRequest<UpdateCommentCommandResponse>
 {
-    public int Id { get; set; }
-    public int UserId { get; set; }
-    public int? ParentId { get; set; }
+    public long Id { get; set; }
+    public long UserId { get; set; }
+    public long? ParentId { get; set; }
     public string Content { get; set; }
     public DateTime CreatedDate { get; set; }
