+++ src/studyCloud/Application/Features/Comments/Comment/Queries/GetByPostIdComment/GetByPostIdCommentQueryRequest.cs
@@ -4,5 +4,5 @@ namespace Application.Features.Comments.Comment.Queries.GetByPostIdComment;
 
 public class GetByPostIdCommentQueryRequest:IRequest<List<GetByPostIdCommentQueryResponse>>
 {
-    public int PostId { get; set; }
+    public long PostId { get; set; }
 }
