+++ src/studyCloud/Application/Features/Comments/CommentFile/Commands/UpdateCommentFile/UpdateCommentFileCommandRequest.cs
@@ -4,8 +4,8 @@ namespace Application.Features.Comments.CommentFile.Commands.UpdateCommentFile;
 
 public class UpdateCommentFileCommandRequest:IRequest<UpdateCommentFileCommandResponse>
 {
-    public int Id { get; set; }
-    public int CommentId { get; set; }
+    public long Id { get; set; }
+    public long CommentId { get; set; }
     public string ImagePath { get; set; }
     public DateTime CreateDate { get; set; }
 }
