+++ src/studyCloud/Application/Features/Comments/CommentLike/Commands/LikeComment/LikeCommentCommandRequest.cs
@@ -4,6 +4,6 @@ namespace Application.Features.Comments.CommentLike.Commands.LikeComment;
 
 public class LikeCommentCommandRequest:IRequest<LikeCommentCommandResponse>
 {
-    public int UserId { get; set; }
-    public int CommentId { get; set; }
+    public long UserId { get; set; }
+    public long CommentId { get; set; }
 }
