+++ src/studyCloud/Application/Features/Comments/CommentLike/Queries/GetUsersLikedComment/GetUsersLikedCommentQueryRequest.cs
@@ -4,5 +4,5 @@ namespace Application.Features.Comments.CommentLike.Queries.GetUsersLikedComment
 
 public class GetUsersLikedCommentQueryRequest:IRequest<List<GetUsersLikedCommentQueryResponse>>
 {
-    public int CommentId { get; set; }
+    public long CommentId { get; set; }
 }
