+++ src/studyCloud/Application/Features/Feeds/PostImageFile/Commands/DeletePostImage/DeletePostFileCommandResponse.cs
@@ -2,6 +2,6 @@ namespace Application.Features.Feeds.PostImageFile.Commands.DeletePostImage;
 
 public class DeletePostFileCommandResponse
 {
-    public int Id { get; set; }
-    public int ImageId { get; set; }
+    public long Id { get; set; }
+    public long ImageId { get; set; }
 }
