+++ src/studyCloud/Application/Features/Feeds/PostImageFile/Commands/DeletePostImage/DeletePostImageCommandRequest.cs
@@ -4,5 +4,5 @@ namespace Application.Features.Feeds.PostImageFile.Commands.DeletePostImage;
 
 public class DeletePostImageCommandRequest:IRequest<DeletePostFileCommandResponse>
 {
-    public int ImageId { get; set; }
+    public long ImageId { get; set; }
 }
