+++ src/studyCloud/Application/Features/Feeds/PostLike/Commands/LikePost/LikePostCommandRequest.cs
@@ -4,6 +4,6 @@ namespace Application.Features.Feeds.PostLike.Commands.LikePost;
 
 public class LikePostCommandRequest:IRequest<LikePostCommandResponse>
 {
-    public int UserId { get; set; }
-    public int PostId { get; set; }
+    public long UserId { get; set; }
+    public long PostId { get; set; }
 }
