+++ src/studyCloud/Application/Features/Feeds/PostLike/Commands/LikePost/LikePostCommandResponse.cs
@@ -2,7 +2,7 @@ namespace Application.Features.Feeds.PostLike.Commands.LikePost;
 
 public class LikePostCommandResponse
 {
-    public int Id { get; set; }
-    public int UserId { get; set; }
-    public int PostId { get; set; }
+    public long Id { get; set; }
+    public long UserId { get; set; }
+    public long PostId { get; set; }
 }
