+++ src/studyCloud/Application/Features/Feeds/PostLike/Commands/UnLikePost/UnLikePostCommandRequest.cs
@@ -4,6 +4,6 @@ namespace Application.Features.Feeds.PostLike.Commands.UnLikePost;
 
 public class UnLikePostCommandRequest:IRequest<UnLikePostCommandResponse>
 {
-   public int UserId { get; set; }
-   public int PostId { get; set; }
+   public long UserId { get; set; }
+   public long PostId { get; set; }
 }
