+++ src/studyCloud/Application/Features/Follows/Commands/CreateFollow/CreateFollowCommandRequest.cs
@@ -4,6 +4,6 @@ namespace Application.Features.Follows.Commands.CreateFollow;
 
 public class CreateFollowCommandRequest:IRequest<CreateFollowCommandResponse>
 {
-    public int? FollowerId { get; set; }
-    public int? FollowingId { get; set; }
+    public long? FollowerId { get; set; }
+    public long? FollowingId { get; set; }
 }
