+++ src/studyCloud/Application/Features/Follows/Commands/UnFollow/UnFollowCommandRequest.cs
@@ -4,6 +4,6 @@ namespace Application.Features.Follows.Commands.UnFollow;
 
 public class UnFollowCommandRequest:IRequest<UnFollowCommandResponse>
 {
-    public int FollowerId { get; set; }
-    public int? FollowingId { get; set; }
+    public long FollowerId { get; set; }
+    public long? FollowingId { get; set; }
 }
