+++ src/studyCloud/Application/Features/Follows/Commands/UpdateFollow/UpdateFollowCommandRequest.cs
@@ -4,9 +4,9 @@ namespace Application.Features.Follows.Commands.UpdateFollow;
 
 public class UpdateFollowCommandRequest : IRequest<UpdateFollowCommandResponse>
 {
-    public int Id { get; set; }
-    public int FollowerId { get; set; }
-    public int FollowingId { get; set; }
+    public long Id { get; set; }
+    public long FollowerId { get; set; }
+    public long FollowingId { get; set; }
 
     
 }
