+++ src/studyCloud/Application/Features/Lessons/Lesson/Commands/CreateLesson/CreateLessonCommandRequest.cs
@@ -5,5 +5,5 @@ namespace Application.Features.Lessons.Lesson.Commands.CreateLesson;
 public class CreateLessonCommandRequest:IRequest<CreateLessonCommandResponse>
 {
     public string Name { get; set; }
-    public int SubCategoryId { get; set; }
+    public long SubCategoryId { get; set; }
 }
