+++ src/studyCloud/Application/Features/Lessons/Lesson/Commands/UpdateLesson/UpdateLessonCommandRequest.cs
@@ -4,7 +4,7 @@ namespace Application.Features.Lessons.Lesson.Commands.UpdateLesson;
 
 public class UpdateLessonCommandRequest:IRequest<UpdateLessonCommandResponse>
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; }
-    public int SubCategoryId { get; set; }
+    public long SubCategoryId { get; set; }
 }
