+++ src/studyCloud/Application/Features/Lessons/LessonSubject/Commands/CreateLessonSubject/CreateLessonSubjectCommandRequest.cs
@@ -5,6 +5,6 @@ namespace Application.Features.Lessons.LessonSubject.Commands.CreateLessonSubjec
 public class CreateLessonSubjectCommandRequest:IRequest<CreateLessonSubjectCommandResponse>
 {
     public string Name { get; set; }
-    public int LessonId { get; set; }
-    public int? ParentId { get; set; }
+    public long LessonId { get; set; }
+    public long? ParentId { get; set; }
 }
