+++ src/studyCloud/Application/Features/Lessons/LessonSubject/Commands/UpdateLessonSubject/UpdateLessonSubjectCommandRequest.cs
@@ -4,8 +4,8 @@ namespace Application.Features.Lessons.LessonSubject.Commands.UpdateLessonSubjec
 
 public class UpdateLessonSubjectCommandRequest:IRequest<UpdateLessonSubjectCommandResponse>
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; }
-    public int LessonId { get; set; }
-    public int? ParentId { get; set; }
+    public long LessonId { get; set; }
+    public long? ParentId { get; set; }
 }
