+++ src/studyCloud/Application/Features/Mentions/Commands/CreateMention/CreateMentionCommandRequest.cs
@@ -5,10 +5,10 @@ namespace Application.Features.Mentions.Commands.CreateMention;
 
 public class CreateMentionCommandRequest:IRequest<CreateMentionCommandResponse>
 {
-    public int AgentId { get; set; }
-    public int TargetId { get; set; }
-    public int PostId { get; set; }
-    public int CommentId { get; set; }
+    public long AgentId { get; set; }
+    public long TargetId { get; set; }
+    public long PostId { get; set; }
+    public long CommentId { get; set; }
     public MentionType MentionType { get; set; }
