+++ src/studyCloud/Application/Features/Mentions/Commands/CreateMention/CreateMentionCommandResponse.cs
@@ -4,11 +4,11 @@ namespace Application.Features.Mentions.Commands.CreateMention;
 
 public class CreateMentionCommandResponse
 {
-    public int Id { get; set; }
-    public int AgentId { get; set; }
-    public int TargetId { get; set; }
-    public int PostId { get; set; }
-    public int CommentId { get; set; }
+    public long Id { get; set; }
+    public long AgentId { get; set; }
+    public long TargetId { get; set; }
+    public long PostId { get; set; }
+    public long CommentId { get; set; }
     public MentionType MentionType { get; set; }
     public DateTime CreatedDate { get; set; }
 }
