+++ src/studyCloud/Application/Features/Mentions/Commands/UpdateMention/UpdateMentionCommandRequest.cs
@@ -5,11 +5,11 @@ namespace Application.Features.Mentions.Commands.UpdateMention;
 
 public class UpdateMentionCommandRequest : IRequest<UpdateMentionCommandResponse>
 {
-    public int Id { get; set; }
-    public int AgentId { get; set; }
-    public int TargetId { get; set; }
-    public int PostId { get; set; }
-    public int CommentId { get; set; }
+    public long Id { get; set; }
+    public long AgentId { get; set; }
+    public long TargetId { get; set; }
+    public long PostId { get; set; }
+    public long CommentId { get; set; }
     public MentionType MentionType { get; set; }
