+++ src/studyCloud/Application/Features/Quizzes/Answer/Commands/CreateAnswer/CreateAnswerCommandRequest.cs
@@ -4,7 +4,7 @@ namespace Application.Features.Quizzes.Answer.Commands.CreateAnswer;
 
 public class CreateAnswerCommandRequest : IRequest<CreateAnswerCommandResponse>
 {
-    public int QuestionId { get; set; }
+    public long QuestionId { get; set; }
     public string Text { get; set; }
     public bool IsCorrect { get; set; }
