+++ src/studyCloud/Application/Features/Quizzes/Answer/Commands/CreateAnswer/CreateAnswerCommandResponse.cs
@@ -2,8 +2,8 @@ namespace Application.Features.Quizzes.Answer.Commands.CreateAnswer;
 
 public class CreateAnswerCommandResponse
 {
-    public int Id { get; set; }
-    public int QuestionId{ get; set; }
+    public long Id { get; set; }
+    public long QuestionId{ get; set; }
     public string Text { get; set; }
     public bool IsCorrect { get; set; }
 }
