+++ src/studyCloud/Application/Features/Quizzes/Answer/Commands/UpdateAnswer/UpdateAnswerCommandRequest.cs
@@ -4,8 +4,8 @@ namespace Application.Features.Quizzes.Answer.Commands.UpdateAnswer;
 
 public class UpdateAnswerCommandRequest : IRequest<UpdatedAnswerCommandResponse>
 {
-    public int Id { get; set; }
-    public int QuestionId { get; set; }
+    public long Id { get; set; }
+    public long QuestionId { get; set; }
     public string Text { get; set; }
     public bool IsCorrect { get; set; }
