+++ src/studyCloud/Application/Features/Quizzes/Answer/Queries/GetByIdAnswer/GetByIdAnswerQueryResponse.cs
@@ -2,8 +2,8 @@ namespace Application.Features.Quizzes.Answer.Queries.GetByIdAnswer;
 
 public class GetByIdAnswerQueryResponse
 {
-    public int Id { get; set; }
-    public int Question{ get; set; }
+    public long Id { get; set; }
+    public long Question{ get; set; }
     public string Text { get; set; }
     public bool IsCorrect { get; set; }
 }
