+++ src/studyCloud/Application/Features/Quizzes/Question/Commands/CreateQuestion/CreateQuestionCommandResponse.cs
@@ -2,7 +2,7 @@ namespace Application.Features.Quizzes.Question.Commands.CreateQuestion;
 
 public class CreateQuestionCommandResponse
 {
-    public int Id { get; set; }
-    public int QuizId { get; set; }
+    public long Id { get; set; }
+    public long QuizId { get; set; }
     public string ImageUrl { get; set; }
 }
