+++ src/studyCloud/Application/Features/Quizzes/Question/Commands/UpdateQuestion/UpdateQuestionCommandRequest.cs
@@ -4,8 +4,8 @@ namespace Application.Features.Quizzes.Question.Commands.UpdateQuestion;
 
 public class UpdateQuestionCommandRequest : IRequest<UpdateQuestionCommandResponse>
 {
-    public int Id { get; set; }
-    public int QuizId { get; set; }
+    public long Id { get; set; }
+    public long QuizId { get; set; }
     public string Text { get; set; }
