+++ src/studyCloud/Application/Features/Quizzes/Question/Commands/UpdateQuestion/UpdateQuestionCommandResponse.cs
@@ -2,7 +2,7 @@ namespace Application.Features.Quizzes.Question.Commands.UpdateQuestion;
 
 public class UpdateQuestionCommandResponse
 {
-    public int Id { get; set; }
-    public int QuizId { get; set; }
+    public long Id { get; set; }
+    public long QuizId { get; set; }
     public string Text { get; set; }
 }
