+++ src/studyCloud/Application/Features/Quizzes/Question/Queries/GetListQuestion/ListQuestionQueryResponse.cs
@@ -2,8 +2,8 @@ namespace Application.Features.Quizzes.Question.Queries.GetListQuestion;
 
 public class ListQuestionQueryResponse
 {
-    public int Id { get; set; }
-    public int QuizId { get; set; }
+    public long Id { get; set; }
+    public long QuizId { get; set; }
     public string Text { get; set; }
     public List<Domain.Entities.Quizzes.Answer> Answers { get; set; }
 }
