+++ src/studyCloud/Application/Features/Quizzes/Quiz/Commands/UpdateQuiz/UpdateQuizCommandRequest.cs
@@ -4,10 +4,10 @@ namespace Application.Features.Quizzes.Quiz.Commands.UpdateQuiz;
 
 public class UpdateQuizCommandRequest : IRequest<UpdateQuizCommandResponse>
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; }
-    public int CategoryId { get; set; }
-    public int SubCategoryId { get; set; }
+    public long CategoryId { get; set; }
+    public long SubCategoryId { get; set; }
 
    
 }
