+++ src/studyCloud/Application/Features/Quizzes/QuizHistory/Commands/CreateQuizHistory/CreateQuizHistoryCommandRequest.cs
@@ -4,8 +4,6 @@ namespace Application.Features.Quizzes.QuizHistory.Commands.CreateQuizHistory;
 
 public class CreateQuizHistoryCommandRequest : IRequest<CreateQuizHistoryCommandResponse>
 {
-    public int UserId { get; set; }
-    public int QuizId { get; set; }
-
-    
+    public long UserId { get; set; }
+    public long QuizId { get; set; }
 }
