+++ src/studyCloud/Application/Features/Quizzes/QuizHistory/Commands/CreateQuizHistory/CreateQuizHistoryCommandResponse.cs
@@ -2,8 +2,8 @@ namespace Application.Features.Quizzes.QuizHistory.Commands.CreateQuizHistory;
 
 public class CreateQuizHistoryCommandResponse
 {
-    public int Id { get; set; }
-    public int UserId { get; set; }
-    public int QuizId { get; set; }
+    public long Id { get; set; }
+    public long UserId { get; set; }
+    public long QuizId { get; set; }
     public DateTime QuizDate { get; set; }
 }
