+++ src/studyCloud/Application/Features/Quizzes/QuizHistory/Commands/UpdateQuizHistory/UpdateQuizHistoryCommandRequest.cs
@@ -4,9 +4,9 @@ namespace Application.Features.Quizzes.QuizHistory.Commands.UpdateQuizHistory;
 
 public class UpdateQuizHistoryCommandRequest : IRequest<UpdateQuizHistoryQueryResponse>
 {
-    public int Id { get; set; }
-    public int UserId { get; set; }
-    public int QuizId { get; set; }
+    public long Id { get; set; }
+    public long UserId { get; set; }
+    public long QuizId { get; set; }
 
    
 }
