+++ src/studyCloud/Application/Features/Quizzes/QuizHistory/Commands/UpdateQuizHistory/UpdateQuizHistoryQueryResponse.cs
@@ -2,8 +2,8 @@ namespace Application.Features.Quizzes.QuizHistory.Commands.UpdateQuizHistory;
 
 public class UpdateQuizHistoryQueryResponse
 {
-    public int Id { get; set; }
-    public int UserId { get; set; }
-    public int QuizId { get; set; }
+    public long Id { get; set; }
+    public long UserId { get; set; }
+    public long QuizId { get; set; }
     public DateTime QuizDate { get; set; }
 }
