+++ src/studyCloud/Application/Features/Quizzes/QuizHistory/Queries/GetByIdQuizHistory/GetByIdQuizHistoryQueryResponse.cs
@@ -2,8 +2,8 @@ namespace Application.Features.Quizzes.QuizHistory.Queries.GetByIdQuizHistory;
 
 public class GetByIdQuizHistoryQueryResponse
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string UserEmail { get; set; }
-    public int QuizId { get; set; }
+    public long QuizId { get; set; }
     public DateTime QuizDate { get; set; }
 }
