+++ src/studyCloud/Application/Features/Quizzes/SelectedAnswer/Commands/CreateSelectedAnswer/CreateSelectedAnswerCommandResponse.cs
@@ -2,9 +2,9 @@ namespace Application.Features.Quizzes.SelectedAnswer.Commands.CreateSelectedAns
 
 public class CreateSelectedAnswerCommandResponse
 {
-    public int Id { get; set; }
-    public int UserId { get; set; }
-    public int QuestionId { get; set; }
-    public int PossibleAnswerId { get; set; }
-    public int QuizHistoryId { get; set; }
+    public long Id { get; set; }
+    public long UserId { get; set; }
+    public long QuestionId { get; set; }
+    public long PossibleAnswerId { get; set; }
+    public long QuizHistoryId { get; set; }
 }
