+++ src/studyCloud/Application/Features/SubCategories/Commands/CreateSubCategory/CreateSubCategoryCommandRequest.cs
@@ -4,7 +4,7 @@ namespace Application.Features.SubCategories.Commands.CreateSubCategory;
 
 public class CreateSubCategoryCommandRequest:IRequest<CreateSubCategoryCommandResponse>
 {
-    public int CategoryId { get; set; }
-    public int? ParentId { get; set; }
+    public long CategoryId { get; set; }
+    public long? ParentId { get; set; }
     public string Name { get; set; }
 }
